app/app/requirements.txt
argh
dateutils
django-cleanup
django-constance[redis]
django-environ
django-extensions
Django>=3.0,<4.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
huey
psycopg2-binary
python-dotenv
pyyaml
supervisor
watchdog
